Add visualization thumbnail
