+++ .gitignore
@@ -122,6 +122,8 @@ docs/source/reference/visualization/generated/
 docs/source/reference/samplers/generated
 docs/source/sg_execution_times.rst
 docs/source/tutorial/**
+docs/source/auto_visualization_examples
+docs/source/auto_visualization_matplotlib_examples
 !docs/source/tutorial/index.rst
 
 # asv
