+++ docs/Makefile
@@ -28,10 +28,13 @@ clean:
 	rm -rf $(BUILDDIR)/*
 	rm -rf source/reference/generated/
 	rm -rf source/reference/visualization/generated/
-	rm source/sg_execution_times.rst
+	rm -f source/sg_execution_times.rst
 	rm -rf source/tutorial/10_key_features
 	rm -rf source/tutorial/20_recipes
 	rm -f ../tutorial/**/*.db
 	rm -rf ../tutorial/20_recipes/artifacts
+	rm -f ../tutorial/20_recipes/journal.log
 	rm -rf ../tutorial/20_recipes/tmp
-	rm -rf ../tutorial/20_recipes/best_atoms.png
+	rm -f ../tutorial/20_recipes/best_atoms.png
+	rm -rf ./source/auto_visualization_matplotlib_examples
+	rm -rf ./source/auto_visualization_examples
