+++ docs/source/reference/visualization/index.rst
@@ -15,23 +15,7 @@ The :mod:`~optuna.visualization` module provides utility functions for plotting 
     .. _JupyterLab: https://github.com/jupyterlab/jupyterlab
     .. _installation guide: https://github.com/plotly/plotly.py#jupyterlab-support
 
-.. autosummary::
-   :toctree: generated/
-   :nosignatures:
-
-   optuna.visualization.plot_contour
-   optuna.visualization.plot_edf
-   optuna.visualization.plot_hypervolume_history
-   optuna.visualization.plot_intermediate_values
-   optuna.visualization.plot_optimization_history
-   optuna.visualization.plot_parallel_coordinate
-   optuna.visualization.plot_param_importances
-   optuna.visualization.plot_pareto_front
-   optuna.visualization.plot_rank
-   optuna.visualization.plot_slice
-   optuna.visualization.plot_terminator_improvement
-   optuna.visualization.plot_timeline
-   optuna.visualization.is_available
+.. include:: ../../auto_visualization_examples/index.rst
 
 .. note::
     The following :mod:`optuna.visualization.matplotlib` module uses Matplotlib as a backend.
