+++ docs/source/reference/visualization/matplotlib.rst
@@ -6,20 +6,4 @@ optuna.visualization.matplotlib
 .. note::
     The following functions use Matplotlib as a backend.
 
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    optuna.visualization.matplotlib.plot_contour
-    optuna.visualization.matplotlib.plot_edf
-    optuna.visualization.matplotlib.plot_hypervolume_history
-    optuna.visualization.matplotlib.plot_intermediate_values
-    optuna.visualization.matplotlib.plot_optimization_history
-    optuna.visualization.matplotlib.plot_parallel_coordinate
-    optuna.visualization.matplotlib.plot_param_importances
-    optuna.visualization.matplotlib.plot_pareto_front
-    optuna.visualization.matplotlib.plot_rank
-    optuna.visualization.matplotlib.plot_slice
-    optuna.visualization.matplotlib.plot_terminator_improvement
-    optuna.visualization.matplotlib.plot_timeline
-    optuna.visualization.matplotlib.is_available
+.. include:: ../../auto_visualization_matplotlib_examples/index.rst
