+++ docs/visualization_examples/README.rst
@@ -1,0 +1,3 @@
+.. _visualization-examples-index:
+
+.. _general_visualization_examples:
