+++ docs/visualization_matplotlib_examples/README.rst
@@ -1,0 +1,3 @@
+.. _visualization-matplotlib-examples-index:
+
+.. _general_visualization_matplotlib_examples:
