+++ optuna/visualization/_edf.py
@@ -57,46 +57,6 @@ def plot_edf(
         The plotted EDF assumes that the value of the objective function is in
         accordance with the uniform distribution over the objective space.
 
-    Example:
-
-        The following code snippet shows how to plot EDF.
-
-        .. plotly::
-
-            import math
-
-            import optuna
-
-
-            def ackley(x, y):
-                a = 20 * math.exp(-0.2 * math.sqrt(0.5 * (x ** 2 + y ** 2)))
-                b = math.exp(0.5 * (math.cos(2 * math.pi * x) + math.cos(2 * math.pi * y)))
-                return -a - b + math.e + 20
-
-
-            def objective(trial, low, high):
-                x = trial.suggest_float("x", low, high)
-                y = trial.suggest_float("y", low, high)
-                return ackley(x, y)
-
-
-            sampler = optuna.samplers.RandomSampler(seed=10)
-
-            # Widest search space.
-            study0 = optuna.create_study(study_name="x=[0,5), y=[0,5)", sampler=sampler)
-            study0.optimize(lambda t: objective(t, 0, 5), n_trials=500)
-
-            # Narrower search space.
-            study1 = optuna.create_study(study_name="x=[0,4), y=[0,4)", sampler=sampler)
-            study1.optimize(lambda t: objective(t, 0, 4), n_trials=500)
-
-            # Narrowest search space but it doesn't include the global optimum point.
-            study2 = optuna.create_study(study_name="x=[1,3), y=[1,3)", sampler=sampler)
-            study2.optimize(lambda t: objective(t, 1, 3), n_trials=500)
-
-            fig = optuna.visualization.plot_edf([study0, study1, study2])
-            fig.show()
-
     Args:
         study:
             A target :class:`~optuna.study.Study` object.
