+++ optuna/visualization/_hypervolume_history.py
@@ -35,31 +35,6 @@ def plot_hypervolume_history(
 ) -> "go.Figure":
     """Plot hypervolume history of all trials in a study.
 
-    Example:
-
-        The following code snippet shows how to plot optimization history.
-
-        .. plotly::
-
-            import optuna
-
-
-            def objective(trial):
-                x = trial.suggest_float("x", 0, 5)
-                y = trial.suggest_float("y", 0, 3)
-
-                v0 = 4 * x ** 2 + 4 * y ** 2
-                v1 = (x - 5) ** 2 + (y - 5) ** 2
-                return v0, v1
-
-
-            study = optuna.create_study(directions=["minimize", "minimize"])
-            study.optimize(objective, n_trials=50)
-
-            reference_point=[100., 50.]
-            fig = optuna.visualization.plot_hypervolume_history(study, reference_point)
-            fig.show()
-
     Args:
         study:
             A :class:`~optuna.study.Study` object whose trials are plotted for their hypervolumes.
