+++ optuna/visualization/_optimization_history.py
@@ -178,28 +178,6 @@ def plot_optimization_history(
 ) -> "go.Figure":
     """Plot optimization history of all trials in a study.
 
-    Example:
-
-        The following code snippet shows how to plot optimization history.
-
-        .. plotly::
-
-            import optuna
-
-
-            def objective(trial):
-                x = trial.suggest_float("x", -100, 100)
-                y = trial.suggest_categorical("y", [-1, 0, 1])
-                return x ** 2 + y
-
-
-            sampler = optuna.samplers.TPESampler(seed=10)
-            study = optuna.create_study(sampler=sampler)
-            study.optimize(objective, n_trials=10)
-
-            fig = optuna.visualization.plot_optimization_history(study)
-            fig.show()
-
     Args:
         study:
             A :class:`~optuna.study.Study` object whose trials are plotted for their target values.
