+++ optuna/visualization/_timeline.py
@@ -33,37 +33,6 @@ class _TimelineInfo(NamedTuple):
 def plot_timeline(study: Study) -> "go.Figure":
     """Plot the timeline of a study.
 
-    Example:
-
-        The following code snippet shows how to plot the timeline of a study.
-        Timeline plot can visualize trials with overlapping execution time
-        (e.g., in distributed environments).
-
-        .. plotly::
-
-            import time
-
-            import optuna
-
-
-            def objective(trial):
-                x = trial.suggest_float("x", 0, 1)
-                time.sleep(x * 0.1)
-                if x > 0.8:
-                    raise ValueError()
-                if x > 0.4:
-                    raise optuna.TrialPruned()
-                return x ** 2
-
-
-            study = optuna.create_study(direction="minimize")
-            study.optimize(
-                objective, n_trials=50, n_jobs=2, catch=(ValueError,)
-            )
-
-            fig = optuna.visualization.plot_timeline(study)
-            fig.show()
-
     Args:
         study:
             A :class:`~optuna.study.Study` object whose trials are plotted with
