+++ optuna/visualization/matplotlib/_contour.py
@@ -45,34 +45,10 @@ def plot_contour(
 
     Note that, if a parameter contains missing values, a trial with missing values is not plotted.
 
-    .. seealso::
-        Please refer to :func:`optuna.visualization.plot_contour` for an example.
-
     Warnings:
         Output figures of this Matplotlib-based
         :func:`~optuna.visualization.matplotlib.plot_contour` function would be different from
         those of the Plotly-based :func:`~optuna.visualization.plot_contour`.
-
-    Example:
-
-        The following code snippet shows how to plot the parameter relationship as contour plot.
-
-        .. plot::
-
-            import optuna
-
-
-            def objective(trial):
-                x = trial.suggest_float("x", -100, 100)
-                y = trial.suggest_categorical("y", [-1, 0, 1])
-                return x ** 2 + y
-
-
-            sampler = optuna.samplers.TPESampler(seed=10)
-            study = optuna.create_study(sampler=sampler)
-            study.optimize(objective, n_trials=30)
-
-            optuna.visualization.matplotlib.plot_contour(study, params=["x", "y"])
 
     Args:
         study:
