+++ optuna/visualization/matplotlib/_edf.py
@@ -29,55 +29,6 @@ def plot_edf(
 
     Note that only the complete trials are considered when plotting the EDF.
 
-    .. seealso::
-        Please refer to :func:`optuna.visualization.plot_edf` for an example,
-        where this function can be replaced with it.
-
-    .. note::
-
-        Please refer to `matplotlib.pyplot.legend
-        <https://matplotlib.org/stable/api/_as_gen/matplotlib.pyplot.legend.html>`_
-        to adjust the style of the generated legend.
-
-    Example:
-
-        The following code snippet shows how to plot EDF.
-
-        .. plot::
-
-            import math
-
-            import optuna
-
-
-            def ackley(x, y):
-                a = 20 * math.exp(-0.2 * math.sqrt(0.5 * (x ** 2 + y ** 2)))
-                b = math.exp(0.5 * (math.cos(2 * math.pi * x) + math.cos(2 * math.pi * y)))
-                return -a - b + math.e + 20
-
-
-            def objective(trial, low, high):
-                x = trial.suggest_float("x", low, high)
-                y = trial.suggest_float("y", low, high)
-                return ackley(x, y)
-
-
-            sampler = optuna.samplers.RandomSampler(seed=10)
-
-            # Widest search space.
-            study0 = optuna.create_study(study_name="x=[0,5), y=[0,5)", sampler=sampler)
-            study0.optimize(lambda t: objective(t, 0, 5), n_trials=500)
-
-            # Narrower search space.
-            study1 = optuna.create_study(study_name="x=[0,4), y=[0,4)", sampler=sampler)
-            study1.optimize(lambda t: objective(t, 0, 4), n_trials=500)
-
-            # Narrowest search space but it doesn't include the global optimum point.
-            study2 = optuna.create_study(study_name="x=[1,3), y=[1,3)", sampler=sampler)
-            study2.optimize(lambda t: objective(t, 1, 3), n_trials=500)
-
-            optuna.visualization.matplotlib.plot_edf([study0, study1, study2])
-
     Args:
         study:
             A target :class:`~optuna.study.Study` object.
