+++ optuna/visualization/matplotlib/_hypervolume_history.py
@@ -23,36 +23,6 @@ def plot_hypervolume_history(
 ) -> "Axes":
     """Plot hypervolume history of all trials in a study with Matplotlib.
 
-    Example:
-
-        The following code snippet shows how to plot optimization history.
-
-        .. plot::
-
-            import optuna
-            import matplotlib.pyplot as plt
-
-
-            def objective(trial):
-                x = trial.suggest_float("x", 0, 5)
-                y = trial.suggest_float("y", 0, 3)
-
-                v0 = 4 * x ** 2 + 4 * y ** 2
-                v1 = (x - 5) ** 2 + (y - 5) ** 2
-                return v0, v1
-
-
-            study = optuna.create_study(directions=["minimize", "minimize"])
-            study.optimize(objective, n_trials=50)
-
-            reference_point=[100, 50]
-            optuna.visualization.matplotlib.plot_hypervolume_history(study, reference_point)
-            plt.tight_layout()
-
-        .. note::
-            You need to adjust the size of the plot by yourself using ``plt.tight_layout()`` or
-            ``plt.savefig(IMAGE_NAME, bbox_inches='tight')``.
-
     Args:
         study:
             A :class:`~optuna.study.Study` object whose trials are plotted for their hypervolumes.
