+++ optuna/visualization/matplotlib/_slice.py
@@ -33,30 +33,6 @@ def plot_slice(
 ) -> "Axes":
     """Plot the parameter relationship as slice plot in a study with Matplotlib.
 
-    .. seealso::
-        Please refer to :func:`optuna.visualization.plot_slice` for an example.
-
-    Example:
-
-        The following code snippet shows how to plot the parameter relationship as slice plot.
-
-        .. plot::
-
-            import optuna
-
-
-            def objective(trial):
-                x = trial.suggest_float("x", -100, 100)
-                y = trial.suggest_categorical("y", [-1, 0, 1])
-                return x ** 2 + y
-
-
-            sampler = optuna.samplers.TPESampler(seed=10)
-            study = optuna.create_study(sampler=sampler)
-            study.optimize(objective, n_trials=10)
-
-            optuna.visualization.matplotlib.plot_slice(study, params=["x", "y"])
-
     Args:
         study:
             A :class:`~optuna.study.Study` object whose trials are plotted for their target values.
